% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myic.R
\name{myic}
\alias{myic}
\title{MYIC}
\usage{
myic(x)
}
\arguments{
\item{x}{(a sample vector)}
}
\value{
a 95% confidence interval
}
\description{
MYIC
}
\examples{
\dontrun myic(c(5.0581, 4.9707, 5.0893, 4.9334, 4.9777, 5.0285, 4.8555, 4.9565, 4.9769, 4.9722, 4.999, 4.9925, 4.9686, 5.0662, 4.9239, 4.9781, 5.0485, 5.0014, 4.9957, 5.0195, 5.0118, 4.9928, 5.0361, 5.0185, 4.9879))
}
