% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{MYNCURVE}
\usage{
myncurve(Y, mu, sigma)
}
\arguments{
\item{A}{value for a variable distributed normally, a mean, and a standard deviation.}
}
\value{
A density curve, an area, and a probability.
}
\description{
MYNCURVE
}
\examples{
\dontrun myncurve(6,10,5)
}
