% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yourbin.R
\name{yourbin}
\alias{yourbin}
\title{#' YOURBIN}
\usage{
yourbin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{An}{iteration count, a sample size, and a probability.}
}
\value{
A table and barplot for a binomial distribution.
}
\description{
#' YOURBIN
}
\examples{
\dontrun yourbin(iter=1000,n=18, p=0.3)
}
