% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yourboot.R
\name{yourboot}
\alias{yourboot}
\title{YOURBOOT}
\usage{
yourboot(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{(iteration count)}

\item{x}{(a sample vector)}

\item{fun}{(measure of central tendency)}

\item{alpha}{(fraction outside of confidence interval)}

\item{cx}{(text expansion factor)}

\item{...}{}
}
\value{
a histogram w/ point and interval estimates
}
\description{
YOURBOOT
}
\examples{
\dontrun yourboot(iter=10000,rnorm(20,mean=10,sd=4))
}
