% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yourclt.R
\name{yourclt}
\alias{yourclt}
\title{YOURCLT}
\usage{
yourclt(n = 10, iter = 100, a = 0, b = 5)
}
\arguments{
\item{n}{(a size)}

\item{iter}{(iteration count)}

\item{a}{(a lower limit)}

\item{b}{(an upper limit)}
}
\value{
a histogram and a summation of "iter" columns
}
\description{
YOURCLT
}
\examples{
\dontrun yourclt(n=10,iter=10000)
}
