% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yourmlbeta.R
\name{yourmlbeta}
\alias{yourmlbeta}
\title{YOURMLBETA}
\usage{
yourmlbeta(x, alpha, beta, ...)
}
\arguments{
\item{x}{(a random beta sample)}

\item{...}{}
}
\value{
plot and maximum estimates for alpha and beta
}
\description{
YOURMLBETA
}
\examples{
\dontrun yourmlbeta(x=rbeta(100,shape1=2,shape2=5),alpha=seq(1,4,length=100),beta=seq(2,8,length=100),lwd=2,labcex=1)
}
