% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yourncurve.R
\name{yourncurve}
\alias{yourncurve}
\title{YOURNCURVE}
\usage{
yourncurve(Y, mu, sigma)
}
\arguments{
\item{A}{value for a variable distributed normally, a mean, and a standard deviation.}
}
\value{
A density curve, an area, and a probability.
}
\description{
YOURNCURVE
}
\examples{
\dontrun yourncurve(6,10,5)
}
