% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{zscore}
\alias{zscore}
\title{ZSCORE}
\usage{
zscore(x)
}
\arguments{
\item{A}{numeric vector of values.}
}
\value{
A Z-score quotient.
}
\description{
ZSCORE
}
\examples{
\dontrun zscore(c(1.1,1.2,1.3,1.4,1.5))
}
